step() now works using double instead of long.
Our time values are normalized to doubles, with 1.0 meaning one second
proper, if the game is running.

+++ src/com/avona/games/towerdefence/StationaryObject.java
@@ -16,6 +16,7 @@ public abstract class StationaryObject {
 	 * Update all state of the object based on the changed in-game time.
 	 * 
 	 * @param dt the time delta since the last invocation of step
+	 *   (1.0 = 1s)
 	 */
-	public abstract void step(long dt);
+	public abstract void step(double dt);
 }
